🚚 moved `CoreMode` to `reg_types.py`
